Separated class to store amore scores from class to run amore rotation search
